rename base config to javascript
