+++ configs/javascript.js
